v0.5

+++ ships.h
@@ -1,23 +1,21 @@
-#include<stdio.h>
-#include<stdlib.h>
-#include<string.h>
-#include<stdbool.h>
-#include "objects.h"
 #ifndef SHIPS_H
 #define SHIPS_H
+
+#include<stdbool.h>
+#include "objects.h"
 
 struct Ship{
     struct Slots* loadout [5];
     char* name;
     float speed;
-    float ship_agility;
-    float turret_agility;
+    int ship_agility;
+    int turret_agility;
     unsigned int gear_score;
     unsigned int hit_points;
     unsigned int defence;
 };
 
-struct Ship* initialiseShip(char* ship_name, float speed, float ship_agility, float turet_agility, unsigned int hit_points, unsigned int defence);
+struct Ship* initialiseShip(char* ship_name, float speed, int ship_agility, int turet_agility, unsigned int hit_points, unsigned int defence);
 void setUpShip(struct Ship* s, unsigned int gear[5][2]);
 
 bool addOject(struct Ship* s, struct Object* o);
